update_permission

+++ app/Http/Controllers/Backend/CategoryNewsController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\CategoryNewsRequest;
 use App\Models\AllCategoryTranslite;
 use App\Models\CategoryNews;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Str;
 
@@ -18,13 +19,16 @@ class CategoryNewsController extends BaseController
      */
     public function index()
     {
+        if (!Auth::user()->can('kategori-berita-view')) {
+            abort(403);
+        }
         try {
             $lang = 'id';
-            $data = AllCategoryTranslite::whereHas('CategoryNews')->where('lang',$lang)->paginate(10);
-            return $this->makeView('backend.pages.management.category.news.index',compact('data'));
+            $data = AllCategoryTranslite::whereHas('CategoryNews')->where('lang', $lang)->paginate(10);
+            return $this->makeView('backend.pages.management.category.news.index', compact('data'));
         } catch (\Throwable $th) {
             dd($th);
-            return redirect()->back()->with('error','galga melakukan aksi');
+            return redirect()->back()->with('error', 'galga melakukan aksi');
         }
     }
 
@@ -34,7 +38,10 @@ class CategoryNewsController extends BaseController
     public function create()
     {
         //
-         return $this->makeView('backend.pages.management.category.news.create');
+        if (!Auth::user()->can('kategori-berita-create')) {
+            abort(403);
+        }
+        return $this->makeView('backend.pages.management.category.news.create');
     }
 
     /**
@@ -42,38 +49,41 @@ class CategoryNewsController extends BaseController
      */
     public function store(CategoryNewsRequest $request)
     {
+        if (!Auth::user()->can('kategori-berita-create')) {
+            abort(403);
+        }
         DB::beginTransaction();
         try {
-            if($request->hasFile('images')){
+            if ($request->hasFile('images')) {
                 $path = public_path('assets/images/category/news');
-                $nameImages = time().'.'.$request->images->extension();
+                $nameImages = time() . '.' . $request->images->extension();
                 $request['images'] = $nameImages;
-                $request->images->move($path,$nameImages);
+                $request->images->move($path, $nameImages);
                 $request['images'] = $nameImages;
             }
             $request['slug'] = Str::slug($request->title);
-            $category = New CategoryNews($request->input());
+            $category = new CategoryNews($request->input());
             $category->save();
             $lang_id = new AllCategoryTranslite([
                 'id_category_news' => $category->id,
                 'lang'                  => 'id',
                 'title'               => $request->title,
-                'slug'                => Str::slug($request->title).'-'.$category->id,
+                'slug'                => Str::slug($request->title) . '-' . $category->id,
             ]);
             $lang_id->save();
             $lang_en = new AllCategoryTranslite([
                 'id_category_news' => $category->id,
                 'lang'                  => 'en',
                 'title'               => $request->title_translite,
-                'slug'                => Str::slug($request->title_translite).'-'.$category->id,
+                'slug'                => Str::slug($request->title_translite) . '-' . $category->id,
             ]);
             $lang_en->save();
             DB::commit();
-            return redirect()->route('category-news.index')->with('success','Success Saving Data');
+            return redirect()->route('category-news.index')->with('success', 'Success Saving Data');
         } catch (\Throwable $th) {
             //throw $th;
             dd($th);
-            return redirect()->back()->with('error','Error Action');
+            return redirect()->back()->with('error', 'Error Action');
         }
     }
 
@@ -91,16 +101,19 @@ class CategoryNewsController extends BaseController
     public function edit(string $id)
     {
 
+        if (!Auth::user()->can('kategori-berita-edit')) {
+            abort(403);
+        }
         //
         try {
             // $data = CategoryNews::findOrFail($id);
             $data = CategoryNews::with('transLite')->findOrFail($id);
-            $contentID = $data->transLite->firstWhere('lang','id') ?? '';
-            $contentEN = $data->transLite->firstWhere('lang','en') ?? '';
-            return $this->makeView('backend.pages.management.category.news.edit',compact('data','contentID','contentEN'));
+            $contentID = $data->transLite->firstWhere('lang', 'id') ?? '';
+            $contentEN = $data->transLite->firstWhere('lang', 'en') ?? '';
+            return $this->makeView('backend.pages.management.category.news.edit', compact('data', 'contentID', 'contentEN'));
         } catch (\Throwable $th) {
             dd($th);
-            return redirect()->back()->with('error','Error Action');
+            return redirect()->back()->with('error', 'Error Action');
         }
     }
 
@@ -109,39 +122,42 @@ class CategoryNewsController extends BaseController
      */
     public function update(CategoryNewsRequest $request, string $id)
     {
+        if (!Auth::user()->can('kategori-berita-edit')) {
+            abort(403);
+        }
         DB::beginTransaction();
         try {
             $category = CategoryNews::with('translite')->findOrFail($id);
-            if($request->hasFile('images')){
+            if ($request->hasFile('images')) {
                 $path = public_path('assets/images/category/news/');
-                $old_images = $path.$category->images;
-                if(file_exists($old_images)){
+                $old_images = $path . $category->images;
+                if (file_exists($old_images)) {
                     unlink($old_images);
                 }
-                $nameImages = time().'.'.$request->images->extension();
+                $nameImages = time() . '.' . $request->images->extension();
                 $request['images'] = $nameImages;
-                $request->images->move($path,$nameImages);
+                $request->images->move($path, $nameImages);
                 $request['images'] = $nameImages;
-            }else{
+            } else {
                 $request['images'] = $category->images;
             }
             $request['slug'] = Str::slug($request->title);
             $category->update($request->input());
-            $lang_en = AllCategoryTranslite::where('lang','en')->where('id_category_news',$id)->first();
-            $lang_id = AllCategoryTranslite::where('lang','id')->where('id_category_news',$id)->first();
+            $lang_en = AllCategoryTranslite::where('lang', 'en')->where('id_category_news', $id)->first();
+            $lang_id = AllCategoryTranslite::where('lang', 'id')->where('id_category_news', $id)->first();
             $lang_en->update([
                 'title' => $request->title_translite,
-                'slug' => Str::slug($request->title_translite).'-'.$lang_en->id
+                'slug' => Str::slug($request->title_translite) . '-' . $lang_en->id
             ]);
             $lang_id->update([
                 'title' => $request->title,
-                'slug' => Str::slug($request->title).'-'.$lang_id->id
+                'slug' => Str::slug($request->title) . '-' . $lang_id->id
             ]);
             DB::commit();
-            return redirect()->route('category-news.index')->with('success','Success Updating Data');
+            return redirect()->route('category-news.index')->with('success', 'Success Updating Data');
         } catch (\Throwable $th) {
             dd($th);
-            return redirect()->back()->with('error','Error Action');
+            return redirect()->back()->with('error', 'Error Action');
         }
     }
 
@@ -150,17 +166,20 @@ class CategoryNewsController extends BaseController
      */
     public function destroy($id)
     {
+        if(!Auth::user()->can('kategori-berita-delete')){
+            abort(403);
+        }
         try {
-           $category = CategoryNews::with('transLite')->findOrFail($id);
-           if($category->transLite->isNotEmpty()){
-               $delete = $category->transLite()->delete(); // Akan menghapus semua data terkait
-           }
-           $category->delete();
-           return redirect()->route('category-news.index')->with('success','Success Delete Data');
+            $category = CategoryNews::with('transLite')->findOrFail($id);
+            if ($category->transLite->isNotEmpty()) {
+                $delete = $category->transLite()->delete(); // Akan menghapus semua data terkait
+            }
+            $category->delete();
+            return redirect()->route('category-news.index')->with('success', 'Success Delete Data');
         } catch (\Throwable $th) {
             //throw $th;
             dd($th);
-            return redirect()->back()->with('error','Error Action');
+            return redirect()->back()->with('error', 'Error Action');
         }
     }
 }
